backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
torch
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
PyJWT==2.8.0
Werkzeug==2.3.7
